requirements.txt
cairocffi==0.8.0
CairoSVG==2.0.3
cffi==1.11.0
cssselect==1.0.1
cssselect2==0.2.0
Django==1.11.5
django-autoslug==1.9.3
django-bootstrap4==0.0.3
django-ckeditor==5.3.0
django-debug-toolbar==1.8
django-easy-pdf==0.1.1
django-js-asset==0.1.1
html5lib==0.999999999
httplib2==0.10.3
lxml==4.0.0
olefile==0.44
Pillow==4.2.1
pycparser==2.18
PyPDF2==1.26.0
Pyphen==0.9.4
pytz==2017.2
reportlab==3.4.0
six==1.11.0
sqlparse==0.2.4
tinycss==0.4
tinycss2==0.6.0
WeasyPrint==0.40
webencodings==0.5.1
xhtml2pdf==0.2b1
